Atualizando projeto 26/03/2025 21:07
Acrescentado ProfissionalDeSaude e tbm revisto o status contulta que está enviando null ajustado, agora o idPaciente está recebendo null verificar os valores e nomes

+++ src/main/java/com/SSHSS/Repository/ProfissionalDeSaudeRepository.java
@@ -3,5 +3,5 @@ package com.SSHSS.Repository;
 import com.SSHSS.Model.ProfissionalDeSaude;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProfissionalDeSaudeRepository extends JpaRepository<ProfissionalDeSaude, Integer> {
+public interface ProfissionalDeSaudeRepository extends JpaRepository<ProfissionalDeSaude, Long> {
 }
